Store latitude and longitude position in the files table.
Note: this is a breaking change. The position TEXT column must be added
to the files table. See format details in filedb.sql.

An API call was also added to fill the position data for files
previously added to the filedb database. For new files this data is
added automatically.